requirements.txt
beautifulsoup4==4.12.3
chromadb==0.5.5
emoji==2.12.1
fastapi==0.114.2
google-auth==2.34.0
grpcio==1.66.1
huggingface-hub==0.24.7
joblib==1.4.2
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langgraph==0.2.21
langgraph-checkpoint==1.0.9
langsmith==0.1.120
lxml==5.3.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.45.0
packaging==24.1
protobuf==4.25.4
pydantic==2.9.1
pydantic_core==2.23.3
pypdf==4.3.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
SQLAlchemy==2.0.34
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
unstructured==0.15.12
unstructured-client==0.25.8
urllib3==2.2.3
uvicorn==0.30.6
websockets==13.0.1
slackclient==2.9.4
python-multipart==0.0.9
loguru==0.7.2
prometheus-fastapi-instrumentator==7.0.0
celery==5.4.0